fromseed/slip/requirements.txt
ed25519
base58
ecdsa
pycryptodome
bip_utils
